aegis/requirements.txt
Flask
flask-sqlalchemy
openai
python-dotenv
supabase
groq